chore: move level mask to extension
